+++ HANGRY/HANGRY/Assets.xcassets/BGColor.colorset/Contents.json
@@ -2,12 +2,12 @@
   "colors" : [
     {
       "color" : {
-        "color-space" : "srgb",
+        "color-space" : "display-p3",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.957",
-          "green" : "0.957",
-          "red" : "0.957"
+          "blue" : "0.925",
+          "green" : "0.925",
+          "red" : "0.925"
         }
       },
       "idiom" : "universal"
