+++ HANGRY/HANGRY/Assets.xcassets/addButton.imageset/Contents.json
@@ -19,5 +19,9 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "auto-scaling" : "auto",
+    "template-rendering-intent" : "original"
   }
 }
