+++ HANGRY/HANGRY/Controllers/HomeViewController.swift
@@ -7,38 +7,5 @@
 
 import Foundation
 import UIKit
-import Combine
 
-enum Section {
-    case informationCollection
-    case warrningList
-    case generate
-}
-class HomeViewController: UIViewController {
-    
-    var collectionView = UICollectionView()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        //setupCollectionView()
-        
-    }
-//
-//    func setupCollectionView() {
-//        collectionView.dataSource = self
-//        collectionView.delegate = self
-//        collectionView.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "SectionCell")
-//    }
-}
-
-//extension HomeViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-//    func numberOfSections(in collectionView: UICollectionView) -> Int { 3 }
-//
-//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int { 4 }
-//
-//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-//        let cell =
-//    }
-//
-//}
+class HomeViewController: UIViewController,
