+++ HANGRY/HANGRY/Controllers/ViewController.swift
@@ -7,16 +7,12 @@
 
 import UIKit
 
-class IngredientsViewController: UIViewController, ingredientDataDelegate, photoDelegate  {
-    
+class ViewController: UIViewController, dataDelegate  {
     
     var ingredients = [Ingredient]()
+    var ing = ["vegie oil","23"]
     
     var tableView = UITableView()
-    
-    var imageArray = [UIImage]()
-    var imageKey = String()
-    var imageStore = ImageStore()
     
     let addImage: UIImageView = {
         let image = UIImageView()
@@ -84,7 +80,7 @@ class IngredientsViewController: UIViewController, ingredientDataDelegate, photo
         self.view.addSubview(tableView)
         tableView.backgroundColor = UIColor(named: "BGColor")
         tableView.separatorColor = .clear
-        tableView.register(IngredientTableViewCell.self, forCellReuseIdentifier: "ingredientCell")
+        tableView.register(TableViewCell.self, forCellReuseIdentifier: "ingredientCell")
         tableView.anchor(top: view.safeAreaLayoutGuide.topAnchor, left: view.leftAnchor,
                          bottom: view.bottomAnchor, right: view.rightAnchor,
                          paddingTop: 20, paddingLeft: 0,
@@ -92,7 +88,6 @@ class IngredientsViewController: UIViewController, ingredientDataDelegate, photo
         tableView.reloadData()
         
     }
-    let queue = DispatchQueue(label: "import image", attributes: [.concurrent])
     
     func saveChanges() {
         print(ingredientDocumentArchiveURL)
@@ -172,12 +167,7 @@ class IngredientsViewController: UIViewController, ingredientDataDelegate, photo
         tableView.reloadData()
     }
     
-    func passPhoto(_ imageKeyStore: String) {
-        self.imageKey = imageKeyStore
-        tableView.reloadData()
-    }
     
-
     
     //MARK: - @objc
     @objc func rightNavItemTapped(_ sender: UIButton){
@@ -188,7 +178,7 @@ class IngredientsViewController: UIViewController, ingredientDataDelegate, photo
 
 
 
-extension IngredientsViewController: UITableViewDataSource, UITableViewDelegate {
+extension ViewController: UITableViewDataSource, UITableViewDelegate {
     
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -197,19 +187,10 @@ extension IngredientsViewController: UITableViewDataSource, UITableViewDelegate 
     
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "ingredientCell", for: indexPath) as! IngredientTableViewCell
+        let cell = tableView.dequeueReusableCell(withIdentifier: "ingredientCell", for: indexPath) as! TableViewCell
         let ingredient = ingredients[indexPath.row]
         cell.ingredientNameLabel.text! = ingredient.name
-        cell.getRemainingTime = ingredient.expiresDate
-        //let image = imageArray[indexPath.row]
-        let image = imageStore.fetchImage(forKey: ingredient.ingredientKey)
-        print(ingredient.ingredientKey)
-        if image != nil {
-            cell.ingredientImage.image = image
-        }
-        
-        
-
+        cell.remainingTimeLabel.text = "\(ingredient.expiresDate)"
         return cell
     }
     
@@ -230,8 +211,6 @@ extension IngredientsViewController: UITableViewDataSource, UITableViewDelegate 
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
             tableView.beginUpdates()
-            let ingredient = ingredients[indexPath.row]
-            imageStore.deleteImage(forKey: ingredient.ingredientKey)
             ingredients.remove(at: indexPath.row)
             saveChanges()
             tableView.deleteRows(at: [indexPath], with: .automatic)
