--- HANGRY/HANGRY/Controllers/recipe/RecipeEntryViewController.swift
@@ -1,19 +1,0 @@
-//
-//  RecipeEntryTableViewController.swift
-//  HANGRY
-//
-//  Created by Dua Almahyani on 12/11/2020.
-//
-
-import UIKit
-
-class RecipeEntryViewController: UIViewController {
-    
-    var delegate: recipeDataDelegate?
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-    }
-
-}
