--- HANGRY/HANGRY/Model/ImageStore.swift
@@ -1,63 +1,0 @@
-//
-//  File.swift
-//  LootLogger
-//
-//  Created by Dua Almahyani on 22/10/2020.
-//  Copyright © 2020 Dua Almahyani. All rights reserved.
-//
-
-import UIKit
-
-class ImageStore {
-    let cache = NSCache<NSString, UIImage>()
-    
-    var imageKey = String()
-    
-    init() {
-        self.imageKey = UUID().uuidString
-    }
-    
-    func setImage(_ image: UIImage, forKey key: String){
-        cache.setObject(image, forKey: key as NSString)
-        
-        let url = imageURL(forKey: key)
-        
-        //ZTurn image into JPEG data
-        if let data = image.jpegData(compressionQuality: 0.5) {
-            try? data.write(to: url)
-        }
-        
-        print(imageURL(forKey: key))
-    }
-    
-    func fetchImage(forKey key: String) -> UIImage? {
-        if let existingImage = cache.object(forKey: key as NSString) {
-            return existingImage
-        }
-        
-        let url = imageURL(forKey: key)
-        
-        guard let imageFromDisk = UIImage(contentsOfFile: url.path) else { return nil }
-        
-        cache.setObject(imageFromDisk, forKey: key as NSString)
-        return imageFromDisk
-    }
-    
-    func deleteImage(forKey key: String) {
-        cache.removeObject(forKey: key as NSString)
-        
-        let url = imageURL(forKey: key)
-        
-        do {
-            try FileManager.default.removeItem(at: url)
-        } catch {
-            print("Error removing the image from disk: \(error)")
-        }
-    }
-    
-    func imageURL(forKey key: String) -> URL {
-        let documentsDirectories = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-        let documentDirectory = documentsDirectories.first!
-        return documentDirectory.appendingPathComponent(key)
-    }
-}
